WAGE: Fix crash in Brownie's Dream
I guess it would crash everywhere else as well, if console window would
be placed the same way. The problem is that console window goes off
screen a little in that game, but copyRectToScreen arguments are not
adjusted to stay within screen area.

This commit adds some checks and adjusts these arguments.