SCUMM: Move detection entries for Steam versions to end of list
For some games the Steam Mac version and Mac CD version have the
same executable name and since we use the executable name to
detect the Steam version it becomes a candidate for the CD version
as well. Moving it at the end of the list ensures that the CD version
entry takes precedence. This fixes bug #7060.