CONFIGURE: Enable plugins support for mingw64
Until now, the configure script was unable to detect mingw64
as a platform that supports building plugins.

I tested this with MSYS2/mingw64 and was able to build all
engines dynamically. I opted not to use any wildcards for this
entry, because I don't know which variants are possible here.