SCI32: Allow access to static script strings from some kArray calls
The original interpreter allows referencing strings as arrays, which we
only allow for SCI32 arrays and strings. In the RAMA demo, the game
scripts reference a string in a game script directly (static script
data), without copying it into a new array. In this commit, only the
functionality used by the RAMA demo has been implemented
(kArrayGetElement and kArrayFree), as it was possible to test them.
Any additional functionality requested will currently error out
wherever lookupArray() is called. This is intentional, as we will
need to check these on a per-call basis