GUI: Suppress Unused Variable Warning For Some Build Configurations.
This fix is not totally clean as it spuriously uses the
serverLabelPosition variable in one case of the preprocessor ifdef
configuration, but with the current structure, this is a simple hack to
fix.

A better solution would be to rewrite some of these functions to remove
some of the preprocessor usage if possible.