ADL: Fix Signed vs. Unsigned Compiler Warnings.
This is a subtle issue associated with the Common::Frac usage. The
frac_t type is signed (int32), but the symbols such as FRAC_ONE are
defined by an enumeration which will default to unsigned int for
members. Unsure if the common code needs changing, but other usages fix
the warning by casting the enumeration values to frac_t so doing this
for now.