TITANIC: Workaround to dispense a cold chicken when the fuse is removed
I'm pretty certain this was the original intention of the dispensor,
as it not only makes sense, but there was code in place in the dispensor
StatusChangeMsg method that handled two different types of dispensing,
depending on whether the fuse was in place. But the code was identical;
given other nastyness and code duplication in the whole dispensor logic,
it's pretty clear that this was what was originally intended. Though
given that the dispensor fuse is present in the fusebox by default, it's
likely no-one ever played around with the dispensor with the fuse removed.