SDL: Use guioptions to decide whether to apply aspect ratio correction or not
This checks whether the current domain has the GUIO_NOASPECT defined or not.
If it is not defined, aspect ratio correction is applied. This replaces testing an
explicit list of resolutions (320x200 and 640x400) which meant some games
that should have had the correction (e.g. some SCI 300x190 and 480x300
games) were excluded.