STARK: Add savegame version handling
The save file header has a version attrubute, as well as each serialized
resource tree. The reason for this is that serialized resource trees
are not necessarily rehydrated into objects during the gameplay.
Thus the save version is not always consistent across the serialized resource
trees in a save file.