SUPERNOVA: aspect ratio correction for newspapers
Criezy's comment on 7d5dd79:
In this case the bool should be true. By convention in engines we default
to 1x scaling for 640x480 and we do not for 320x200. And this means that
if ScummVM is setup to use a 2X scaler in windowed mode, and aspect ratio
correction is on, the window will have the same size for both modes.