TITANIC: Workaround music being lost leaving Arboretum in Winter
When you change the Arboretum to Winter and exit, it changes to
a special FrozenArboretum room with the river and everything frozen.
Both the regular and frozen Arboretum rooms have a GondolierMixer..
when leavin one for the other, it's meant to do a two second fadeout
and then the new room's mixer takes over. This doesn't work for
ScummVM currently - it may be due to the simplicity of the QMixer
implementation. So to work around it for the forseeable future,
I've set the TurnOff times to 0 seconds so the old mixer's sounds
are immediately and properly stopped, letting the new mixer take over.