TITANIC: Workaround Conversation area corruption when dragging items
When dragging an item, the bounds for the area of the screen affected
goes from the dragged item to the blinking cursor when the conversation
tab is active. This rect overlapping the conversation area somehow
starts erasing some of the conversation log. Despite some effort, I
haven't been able to track down the exact cause. So the easiest
workaround is simply to make the entire PET as dirty each frame when
dragging an item, which seems to fix the problem. Though it would be
nice, at some point, if the rooot cause could still be identified.